% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssembleObject.R
\name{AssembleObject}
\alias{AssembleObject}
\alias{AssembleAssay}
\alias{AssembleDimReduc}
\alias{AssembleGraph}
\alias{AssembleImage}
\alias{AssembleNeighbor}
\alias{AssembleSeuratCommand}
\title{Assemble an object from an h5Seurat file}
\usage{
AssembleAssay(assay, file, slots = NULL, cell_idx = NULL, verbose = TRUE)

AssembleDimReduc(reduction, file, cell_idx = NULL, verbose = TRUE)

AssembleGraph(graph, file, cell_idx = NULL, verbose = TRUE)

AssembleImage(image, file, cell_idx = NULL, verbose = TRUE)

AssembleNeighbor(neighbor, file, cell_idx = NULL, verbose = TRUE)

AssembleSeuratCommand(cmd, file, verbose = TRUE)
}
\arguments{
\item{assay, reduction, graph, image, neighbor, cmd}{Name of assay, reduction,
graph, image, neighbor, or command to load}

\item{file}{A connected h5Seurat file to pull the data from}

\item{slots}{Optional vector of assay slots to load, defaults to all slots
present in assay}

\item{verbose}{Show progress updates}
}
\value{
\code{AssembleAssay}: An \code{Assay} object
}
\description{
Assemble an object from an h5Seurat file
}
\keyword{internal}
